requirements.txt
streamlit
opencv-python
Pillow
python-dotenv
google-generativeai
numpy